agriai/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
numpy==1.24.3
Pillow==9.5.0
googletrans==4.0.0-rc1
python-dotenv==1.0.0